+++ wadiz/src/main/java/com/prgrms/wadiz/domain/maker/controller/MakerController.java
@@ -6,8 +6,12 @@ import com.prgrms.wadiz.domain.maker.service.MakerService;
 import com.prgrms.wadiz.global.util.resTemplate.ResponseFactory;
 import com.prgrms.wadiz.global.util.resTemplate.ResponseTemplate;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
 
 @RestController
 @RequestMapping("/makers")
@@ -23,13 +27,13 @@ public class MakerController {
     }
 
     @PutMapping("/{makerId}")
-    public ResponseEntity<ResponseTemplate> updateMaker(Long makerId, MakerUpdateRequestDTO dto) {
+    public ResponseEntity<ResponseTemplate> updateMaker(@PathVariable Long makerId, MakerUpdateRequestDTO dto) {
         makerService.updateMaker(makerId, dto);
         return ResponseEntity.ok(ResponseFactory.getSuccessResult());
     }
 
-    @DeleteMapping("/{makerId")
-    public ResponseEntity<ResponseTemplate> deleteMaker(Long makerId) {
+    @DeleteMapping("/{makerId}")
+    public ResponseEntity<ResponseTemplate> deleteMaker(@PathVariable Long makerId) {
         makerService.deleteMaker(makerId);
         return ResponseEntity.ok(ResponseFactory.getSuccessResult());
     }
