+++ wadiz/src/main/java/com/prgrms/wadiz/domain/maker/dto/request/MakerCreateRequestDTO.java
@@ -1,5 +1,6 @@
 package com.prgrms.wadiz.domain.maker.dto.request;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 
 @Builder
