+++ wadiz/src/main/java/com/prgrms/wadiz/domain/maker/respository/MakerRepository.java
@@ -2,7 +2,6 @@ package com.prgrms.wadiz.domain.maker.respository;
 
 import com.prgrms.wadiz.domain.maker.entity.Maker;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 public interface MakerRepository extends JpaRepository<Maker, Long> {
 }
