+++ wadiz/src/main/java/com/prgrms/wadiz/domain/supporter/dto/response/SupporterResponseDTO.java
@@ -5,11 +5,12 @@ import com.prgrms.wadiz.domain.supporter.entity.Supporter;
 import lombok.Builder;
 
 @Builder
-public record SupporterResponseDTO(String name, String email) {
+public record SupporterResponseDTO(String supporterName,
+                                   String supporterEmail) {
     public static SupporterResponseDTO of(String name, String email) {
         return SupporterResponseDTO.builder()
-                .name(name)
-                .email(email)
+                .supporterName(name)
+                .supporterEmail(email)
                 .build();
     }
 }
